% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rstructBio.R
\name{run_rstructBio}
\alias{run_rstructBio}
\title{Function to run NicheToolBox shinyApp}
\usage{
run_rstructBio()
}
\description{
This function initiates rstructBio GUI
}
\examples{
# Run rstructBio package
# run_rstructBio()
}
